+++ slang/test/integration/main/fallback_base_locale_test.dart
@@ -8,6 +8,7 @@ import 'package:slang/src/builder/model/translation_map.dart';
 import 'package:test/test.dart';
 
 import '../../util/resources_utils.dart';
+import '../../util/setup.dart';
 
 void main() {
   late String compactInput;
@@ -21,6 +22,10 @@ void main() {
   late String specialExpectedMainOutput;
   late String specialExpectedEnOutput;
   late String specialExpectedDeOutput;
+
+  setUpAll(() {
+    runSetupAll();
+  });
 
   setUp(() {
     compactInput = loadResource('main/csv_compact.csv');