+++ slang/test/integration/main/rich_text_test.dart
@@ -6,10 +6,15 @@ import 'package:slang/src/builder/model/translation_map.dart';
 import 'package:test/test.dart';
 
 import '../../util/resources_utils.dart';
+import '../../util/setup.dart';
 
 void main() {
   late String input;
   late String expectedEnOutput;
+
+  setUpAll(() {
+    runSetupAll();
+  });
 
   setUp(() {
     input = loadResource('main/json_rich_text.json');