+++ slang/test/integration/main/yaml_test.dart
@@ -6,6 +6,7 @@ import 'package:slang/src/builder/model/translation_map.dart';
 import 'package:test/test.dart';
 
 import '../../util/resources_utils.dart';
+import '../../util/setup.dart';
 
 void main() {
   late String enInput;
@@ -14,6 +15,10 @@ void main() {
   late String expectedMainOutput;
   late String expectedEnOutput;
   late String expectedDeOutput;
+
+  setUpAll(() {
+    runSetupAll();
+  });
 
   setUp(() {
     enInput = loadResource('main/yaml_en.yaml');